fix: resolve lobby reload hydration mismatch
